Code/requirements.txt
streamlit == 1.22.0
psycopg2-binary==2.9.6
pandas==2.0.1
python-dotenv==1.0.0